feat: flujo auditor completo funcionando

+++ proyectoIntegradorBack/src/main/java/com/co2x/dmrv/repository/ReporteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReporteRepository extends JpaRepository<Reporte, Long> {
+public interface ReporteRepository extends JpaRepository<Reporte, Integer> {
 }
